fixed db issues
- command triggering now works flawlessly
