+++ requirements.txt
@@ -1,7 +1,6 @@
 pyrogram
 tgcrypto
 pymoe
-requests
 aiohttp
 telegraph==1.4.1
 requests==2.24.0
