+++ thebot/modules/dev.py
@@ -3,8 +3,9 @@ import sys
 import os
 import re
 import subprocess
-from io import StringIO
-from thebot import dankbot
+from io import StringIO, BytesIO
+from thebot import dankbot, ALLOWED_USERS
+from thebot.sql import chats_db
 from pyrogram import filters
 
 
@@ -16,7 +17,7 @@ async def aexec(code, client, message):
     return await locals()['__aexec'](client, message)
 
 
-@dankbot.on_message(filters.user(895373440) & filters.command("eval"))
+@dankbot.on_message(filters.user(ALLOWED_USERS) & filters.command("eval"))
 async def evaluate(client, message):
     status_message = await message.reply_text("`Running ...`")
     try:
@@ -67,7 +68,7 @@ async def evaluate(client, message):
 
 
 
-@dankbot.on_message(filters.user(895373440) & filters.command("term"))
+@dankbot.on_message(filters.user(ALLOWED_USERS) & filters.command("term"))
 async def terminal(client, message):
     if len(message.text.split()) == 1:
         await message.reply("Usage: `/term echo owo`")
@@ -124,3 +125,44 @@ async def terminal(client, message):
     else:
         await message.reply("**Output:**\n`No Output`")
 
+
+@dankbot.on_message(filters.user(ALLOWED_USERS) & filters.command("stats", prefixes='/'))
+async def stats_text(_, message):
+    stats = "──「 <b>Current stats</b> 」──\n"
+    stats += f"-> <code>{chats_db.num_users()}</code> users, across <code>{chats_db.num_chats()}</code> chats"
+    await message.reply(stats)
+
+
+@dankbot.on_message(~filters.me & filters.user(ALLOWED_USERS) & filters.command("chats", prefixes='/'))
+async def chat_stats(client, message):
+    all_chats = chats_db.get_all_chats() or []
+    chatfile = 'List of chats.\n0. Chat name | Chat ID | Members count\n'
+    P = 1
+    for chat in all_chats:
+        curr_chat = await client.get_chat(chat.chat_id)
+        bot_member = await curr_chat.get_member(1389308739)
+        chat_members = await client.get_chat_members_count(chat.chat_id)
+        chatfile += "{}. {} | {} | {}\n".format(P, chat.chat_name,
+                                                    chat.chat_id, chat_members)
+        P += 1
+
+    with BytesIO(str.encode(chatfile)) as output:
+        output.name = "chatlist.txt"
+        await message.reply_document(
+            document=output,
+            caption="Here is the list of chats in my database.")
+
+
+@dankbot.on_message(filters.all & filters.group)
+def log_user(client, message):
+    chat = message.chat
+    chats_db.update_user(message.from_user.id, message.from_user.username, chat.id,
+                    chat.title)
+
+    if message.reply_to_message:
+        chats_db.update_user(message.reply_to_message.from_user.id,
+                        message.reply_to_message.from_user.username, chat.id,
+                        chat.title)
+
+    if message.forward_from:
+        chats_db.update_user(message.forward_from.id, message.forward_from.username)
